Generalised object icon
