+++ app/lib/features/activities/widgets/space_activities_section/item_widgets/containers/activity_bigger_visual_container_widget.dart
@@ -1,11 +1,11 @@
 import 'package:acter/common/themes/colors/color_scheme.dart';
+import 'package:acter/features/activities/widgets/space_activities_section/item_widgets/object_icon_widget.dart';
 import 'package:acter/features/comments/widgets/time_ago_widget.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:acter/common/providers/room_providers.dart';
 import 'package:acter_avatar/acter_avatar.dart';
-import 'package:phosphor_flutter/phosphor_flutter.dart';
 
 class ActivityBiggerVisualContainerWidget extends ConsumerWidget {
   final ActivityObject? activityObject;
@@ -115,11 +115,7 @@ class ActivityBiggerVisualContainerWidget extends ConsumerWidget {
           ),
         ),
         if (activityObject != null)
-          Icon(
-            _getActivityObjectIcon(),
-            size: 16,
-            color: colorScheme.surfaceTint,
-          ),
+          ObjectIconWidget(objectType: activityObject?.typeStr()),
       ],
     );
   }
@@ -140,17 +136,5 @@ class ActivityBiggerVisualContainerWidget extends ConsumerWidget {
         ],
       ),
     );
-  }
-
-  IconData _getActivityObjectIcon() {
-    return switch (activityObject?.typeStr()) {
-      'news' => PhosphorIconsRegular.rocketLaunch,
-      'story' => PhosphorIconsRegular.book,
-      'event' => PhosphorIconsRegular.calendar,
-      'pin' => PhosphorIconsRegular.pushPin,
-      'task-list' => PhosphorIconsRegular.listChecks,
-      'task' => PhosphorIconsRegular.checkCircle,
-      _ => PhosphorIconsRegular.question,
-    };
   }
 }
